Renamed logging.py due to an issue I was getting trying to import markdown. Adjusted the space replacement feature in create.py to accommodate more types of spaces. Some other minor changes
