+++ requirements-dev.lock
@@ -18,6 +18,10 @@ coverage==7.4.4
     # via pytest-cov
 emoji==2.10.1
     # via signal-export
+exceptiongroup==1.2.2
+    # via pytest
+importlib-metadata==8.5.0
+    # via markdown
 iniconfig==2.0.0
     # via pytest
 markdown==3.6
@@ -50,9 +54,14 @@ soupsieve==2.5
     # via beautifulsoup4
 sqlcipher3-wheels==0.5.2.post1
     # via signal-export
+tomli==2.2.1
+    # via coverage
+    # via pytest
 typer==0.12.5
     # via signal-export
 types-emoji==2.1.0.3
 types-markdown==3.6.0.20240316
 typing-extensions==4.10.0
     # via typer
+zipp==3.21.0
+    # via importlib-metadata
