+++ requirements.lock
@@ -16,6 +16,8 @@ click==8.1.7
     # via typer
 emoji==2.10.1
     # via signal-export
+importlib-metadata==8.5.0
+    # via markdown
 markdown==3.6
     # via signal-export
 markdown-it-py==3.0.0
@@ -38,3 +40,5 @@ typer==0.12.5
     # via signal-export
 typing-extensions==4.10.0
     # via typer
+zipp==3.21.0
+    # via importlib-metadata
