+++ sigexport/create.py
@@ -1,7 +1,9 @@
+import re
+
 from pathlib import Path
 
 from sigexport import models, utils
-from sigexport.logging import log
+from sigexport.signal_logging import log
 
 
 def create_message(
@@ -47,7 +49,7 @@ def create_message(
     for att in msg.attachments:
         file_name = att["fileName"]
         path = Path("media") / file_name
-        path = Path(str(path).replace(" ", "%20"))
+        path = Path(re.sub(r'\s', '%20', str(path)))
         attachments.append(models.Attachment(name=file_name, path=str(path)))
 
     reactions: list[models.Reaction] = []
